Bugfix/new social flow fix vnk 17 03

+++ lib/remote/services/google/google_cloud_storage_uploader.dart
@@ -196,7 +196,7 @@ class GoogleCloudStorageUploader {
             // Calculate upload progress (0% to 90%)
             final uploadProgress = (sent / total) * 0.9;
             onProgress(uploadProgress);
-            if (_logProgress != (uploadProgress * 100)) {
+            if (_logProgress != (uploadProgress * 100).toInt()) {
               _logProgress = (uploadProgress * 100).toInt();
               debugPrint(
                   'Upload progress: $_logProgress% ($sent/$total bytes)');
